+++ ProjectManager/ProjectManager/ViewController/ViewController+DropDelegate.swift
@@ -1,0 +1,14 @@
+//
+//  ViewController+DropDelegate.swift
+//  ProjectManager
+//
+//  Created by 이영우 on 2021/07/22.
+//
+
+import UIKit
+
+extension ViewController: UITableViewDropDelegate {
+    func tableView(_ tableView: UITableView, performDropWith coordinator: UITableViewDropCoordinator) {
+        return
+    }
+}
