+++ ProjectManager/ProjectManager/ViewController/ViewController.swift
@@ -8,9 +8,9 @@ import UIKit
 
 class ViewController: UIViewController {
 
-    var todoTasks: [Task] = [Task(title: "할일1", content: "할일내용1", deadLine: "2021.05.12")]
-    var doingTasks: [Task] = [Task(title: "하고있는일1", content: "하고있는일내용1", deadLine: "2021.05.13")]
-    var doneTasks: [Task] = [Task(title: "한일1", content: "한일내용1", deadLine: "2021.05.22")]
+    var todoTasks: [Task] = [Task(title: "할일1", content: "할일내용1", deadLine: "2021.05.12", state: "todo")]
+    var doingTasks: [Task] = [Task(title: "하고있는일1", content: "하고있는일내용1", deadLine: "2021.05.13", state: "doing")]
+    var doneTasks: [Task] = [Task(title: "한일1", content: "한일내용1", deadLine: "2021.05.22", state: "done")]
     
     lazy var todoTableView: UITableView = {
         let todoTableView: UITableView = UITableView()
@@ -189,6 +189,8 @@ class ViewController: UIViewController {
         for tableview in [todoTableView, doingTableView, doneTableView] {
             tableview.register(ItemTableViewCell.self, forCellReuseIdentifier: ItemTableViewCell.identifier)
             tableview.dataSource = self
+            tableview.dropDelegate = self
+            tableview.dragDelegate = self
         }
     }
     
@@ -200,38 +202,5 @@ class ViewController: UIViewController {
         addSubViews()
         configureConstraints()
         configureTableViews()
-    }
-}
-
-extension ViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        switch tableView {
-        case todoTableView:
-            return todoTasks.count
-        case doingTableView:
-            return doingTasks.count
-        case doneTableView:
-            return doneTasks.count
-        default:
-            return 0
-        }
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: ItemTableViewCell.identifier) as? ItemTableViewCell else {
-            return UITableViewCell()
-        }
-        
-        switch tableView {
-        case todoTableView:
-            cell.configure(task: todoTasks[indexPath.row])
-        case doingTableView:
-            cell.configure(task: doingTasks[indexPath.row])
-        case doneTableView:
-            cell.configure(task: doneTasks[indexPath.row])
-        default:
-            print("error")
-        }
-        return cell
     }
 }
